workflow_platform/requirements.txt
# Django Workflow Platform Requirements

# Core Django
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2
django-extensions==3.2.3

# Database
psycopg2-binary==2.9.9
redis==5.0.4
django-redis==5.4.0

# API & Authentication
djangorestframework-simplejwt==5.3.1
django-oauth-toolkit==2.3.0
django-allauth==0.63.3
dj-rest-auth==5.1.0

# Async & Background Tasks
celery==5.4.0
django-celery-beat==2.6.0
django-celery-results==2.5.1
kombu==5.5.4

# Monitoring & Performance
django-prometheus==2.4.1

# File Handling
Pillow==10.4.0
django-storages==1.14.4
boto3==1.35.3

# API Documentation
drf-spectacular==0.27.2
drf-spectacular-sidecar==2025.8.1

# Validation & Serialization
marshmallow==3.22.0
jsonschema==4.23.0
pydantic==2.8.2

# HTTP & Requests
requests==2.32.3
httpx==0.27.0
aiohttp==3.10.5

# Email
sendgrid==6.11.0
django-anymail==11.0

# Utils
python-decouple==3.8
pytz==2025.1
python-dateutil==2.9.0.post0
shortuuid==1.0.13

# Encryption & Security
cryptography==42.0.8
PyJWT==2.9.0

# Data Processing
pandas==2.2.2
numpy==1.26.4

# Development
black==24.4.2
flake8==7.1.1
pre-commit==3.8.0
factory-boy==3.3.0
pytest==8.3.2
pytest-django==4.9.0
pytest-cov==5.0.0

# Production
gunicorn==22.0.0
uvicorn==0.30.5
whitenoise==6.7.0
psutil
